% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Data analysis for American Airlines flights data}
\usage{
analyze()
}
\value{
list
}
\description{
The function imports a .csv file from an URL which contains data for all flights that departed NYC (JFK, LGA or EWR) in 2013, taken from the nycflights13 package dataset.
Then, it filters American Airlines flights and it makes a sample that will be used for a plot to check the dependence of arrival delay and distance traveled.
Finally, it calculates the average arrival delay of flights departed from NYC in 2013.
}
\examples{
analyze()

}
