% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{catch_data}
\alias{catch_data}
\title{Import flights data from the nycflights13 package dataset}
\usage{
catch_data()
}
\description{
Data for all flights that departed NYC (JFK, LGA or EWR) in 2013.
They're gathered in a data frame and, for our analysis, we're interested in columns:
'dep_delay, arr_delay': departure and arrival delays in minutes; negative times represent early departures/arrivals.
'carrier': two letter carrier abbreviation.
The first two columns are particularly useful for the plot and through the last one we can filter American Airlines data.
}
