% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightsanalysis.R
\docType{package}
\name{flightanalysis}
\alias{flightanalysis}
\title{Flightsanalysis Package}
\description{
This is a package for the 'Coding for Data Science and Data Management' course
at UNIMI. It analyzes 2013 New York flights data.
}
\author{
lucacerab \email{luca.cerabone@studenti.unimi.it}
}
