% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightsanalysis.R
\docType{package}
\name{flightsanalysis}
\alias{flightsanalysis}
\title{Flightsanalysis Package}
\description{
This is a package for the 'Coding for Data Science and Data Management - R module' course
at UNIMI. It contains 3 functions to analyze 2013 New York flights data:
\itemize{
\item catch(): to retrieve the dataset
\item obs(): to plot the variables arr_delay and dep_delay
\item mean_arrdelay(): to calculate the average arrival delay for American Airlines flights departed from NYC in 2013
}
}
\author{
Luca Cerabone \email{luca.cerabone@studenti.unimi.it}
}
