% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lemmesee}
\alias{lemmesee}
\title{Scatter plot of American Airlines departure and arrival delay}
\usage{
lemmesee()
}
\description{
Plotting American Airlines departure and arrival delays to check if there is some sort of dependence. As we can see as dep_delays increase, also arr_delays increase.
}
\details{
Plotting American Airlines departure and arrival delay
}
