% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{mean_aa}
\alias{mean_aa}
\title{American Airlines mean arrival delay}
\usage{
mean_aa()
}
\description{
The function calculate the mean arrival delay for
American Airlines in 2013
}
\details{
Calculating the mean arrival delay for American Airlines in 2013
}
