% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{obs}
\alias{obs}
\title{American Airlines departure and arrival delays plot}
\usage{
obs()
}
\description{
Scatter plot of American Airlines departure and arrival delays to check if there is dependence.
Once you plot it you can see that, as expected, as departure delay increase, also arrival delay increase.
}
