% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{runEXP}
\alias{runEXP}
\title{Interactive interface}
\usage{
runEXP()
}
\value{
shiny app
}
\description{
Runs an interactive interface where you can explore data used in this package.
The shiny app might need a bit to show the plots.
}
