% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_carrier}
\alias{select_carrier}
\title{Average arrival delay for selected carrier}
\usage{
select_carrier(id)
}
\arguments{
\item{ids}{IATA character}
}
\value{
average delays
}
\description{
Before running this function run the catch() one.
Use as argument of this function the IATA carrier code you can find in the flights data frame in the global environment.
If you don't know the corresponding airline company, search it on google (e.g:'B6 IATA code')
}
\examples{
select_carrier("B6")

}
