% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{catch}
\alias{catch}
\title{Import data and filter American Airlines flights}
\usage{
catch()
}
\value{
data.frame
}
\description{
This function grabs data for all flights that departed NYC (JFK, LGA or EWR) in 2013 and store the data frame to Global Environment.
Data taken from the nycflights13 package dataset.
run 'flights = catch()' if you want to use the select_carrier() function.
}
\examples{
catch()

}
