% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcns.R
\name{catch}
\alias{catch}
\title{Import data and filter American Airlines flights}
\usage{
catch()
}
\description{
Data for all flights that departed NYC (JFK, LGA or EWR) in 2013, taken from the nycflights13 package dataset.
You can look at American Airlines data frame by clicking on it in Global Environment.
For our analysis, we're interested in columns:
\itemize{
\item 'dep_delay, arr_delay': departure and arrival delays in minutes; negative times represent early departures/arrivals.
\item 'carrier': two letter carrier abbreviation.
}
}
