% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{catch}
\alias{catch}
\title{Import flights data}
\usage{
catch()
}
\description{
Data for all flights that departed NYC (JFK, LGA or EWR) in 2013, taken from the nycflights13 package dataset.
For our analysis, we're interested in columns:
dep_delay, arr_delay': departure and arrival delays in minutes; negative times represent early departures/arrivals.
'carrier': two letter carrier abbreviation.
}
