% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{catch}
\alias{catch}
\title{Import data and filter American Airlines flights}
\usage{
catch()
}
\value{
data.frame
}
\description{
This function grabs data for all flights that departed NYC (JFK, LGA or EWR) in 2013.
Data taken from the nycflights13 package dataset.
Run 'flights = catch()' before using the 'select_carrier()' function.
}
\examples{
catch()

}
