% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{catch_data}
\alias{catch_data}
\title{Importing flights data from the nycflights13 dataset}
\usage{
catch_data()
}
\description{
Reading from the csv file and storing the data frame in the variable named 'flights'
}
\details{
Importing data
}
