% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightsanalysis.R
\docType{package}
\name{flightsanalysis}
\alias{flightsanalysis}
\title{Flightsanalysis Package}
\description{
This is a package for the 'Coding for Data Science and Data Management - R module' course
at UNIMI. It contains 4 functions to analyze 2013 New York flights data:
\itemize{
\item analyze(): the main function of the package that analyzes data for American Airlines flights.
\item catch(): to retrieve flights data frame.
\item runEXP(): to launch an interactive interface to explore data.
}
}
\author{
Luca Cerabone \email{luca.cerabone@studenti.unimi.it}
}
