% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lemmesee}
\alias{lemmesee}
\title{American Airlines departure and arrival delays plot}
\usage{
lemmesee()
}
\description{
Scatter plot of American Airlines departure and arrival delays to check if there is some sort of dependence.
Once you plot it you can see that as dep_delays increase, also arr_delays increase.
}
