% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{mean_arrdelay}
\alias{mean_arrdelay}
\title{American Airlines average arrival delay}
\usage{
mean_arrdelay()
}
\description{
The function calculates the average arrival delay for American Airlines flights departed from New York in 2013.
}
