% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{runEXP}
\alias{runEXP}
\title{Interactive interface}
\usage{
runEXP()
}
\value{
shiny app
}
\description{
Runs an interactive interface where you can explore data used in this package. If there are no data for selected input, the plot will be blank.
To speed the process up, you can run first the command 'flights = catch()', if not already done.
}
